+++ src/components/Text/Text.styles.ts
@@ -1,66 +1,8 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-import { palette, semanticColors } from '../../util/colors';
-import { TextProps, TextSize, TextColor } from './Text';
-import { ellipsisStyle, fontWeightBold } from '../../util/styles';
+import { TextProps, TextSize } from './Text';
+import { ellipsisStyle, fontWeightBold, fontSizes, lineHeights, textColors, Sizes } from '../../util/styles/fonts';
 import { IRawStyle } from '@uifabric/styling';
 import { CSSProperties } from 'react';
-
-export type Sizes = { [size in TextSize]: string };
-
-export const textColors: { [color in TextColor]: string | undefined } = {
-  primary: semanticColors.bodyText,
-  secondary: palette.neutralPrimaryAlt,
-  metadata: palette.neutralSecondary,
-  white: palette.white,
-  error: palette.redDark,
-  disabled: semanticColors.disabledBodyText,
-};
-
-export const lineHeights: Sizes = {
-  xSmall: '1.6rem',
-  small: '1.6rem',
-  mediumSub: '2.0rem',
-  medium: '2.0rem',
-  large: '2.4rem',
-  xLarge: '2.8rem',
-  xxLarge: '2.8rem',
-};
-
-export const fontSizes: Sizes = {
-  xSmall: '1.0rem',
-  small: '1.2rem',
-  mediumSub: '1.4rem',
-  medium: '1.5rem',
-  large: '1.8rem',
-  xLarge: '2.2rem',
-  xxLarge: '2.4rem',
-};
-
-const getHeight = (size?: TextSize) => {
-  const heights: Sizes = {
-    xSmall: '1.6rem',
-    small: '1.6rem',
-    mediumSub: '2rem',
-    medium: '2rem',
-    large: '2.4rem',
-    xLarge: '2.8rem',
-    xxLarge: '2.8rem',
-  };
-  return size ? heights[size] : undefined;
-};
-
-const getVerticalAlign = (size?: TextSize) => {
-  const verticalAligns: Sizes = {
-    xSmall: '-0.4rem',
-    small: '-0.3rem',
-    mediumSub: '-0.5rem',
-    medium: '-0.4rem',
-    large: '-0.5rem',
-    xLarge: '-0.5rem',
-    xxLarge: '-0.4rem',
-  };
-  return size ? verticalAligns[size] : undefined;
-};
 
 export const getStyles = (props: TextProps): IRawStyle => {
   const { size, maxWidth, bold, uppercase, color } = props;
@@ -76,6 +18,34 @@ export const getStyles = (props: TextProps): IRawStyle => {
     maxWidth: maxWidth || undefined,
     verticalAlign: maxWidth ? '-0.4rem' : undefined,
   };
+};
+
+const heights: Sizes = {
+  xSmall: '1.6rem',
+  small: '1.6rem',
+  mediumSub: '2rem',
+  medium: '2rem',
+  large: '2.4rem',
+  xLarge: '2.8rem',
+  xxLarge: '2.8rem',
+};
+
+const getHeight = (size?: TextSize) => {
+  return size ? heights[size] : undefined;
+};
+
+const verticalAligns: Sizes = {
+  xSmall: '-0.4rem',
+  small: '-0.3rem',
+  mediumSub: '-0.5rem',
+  medium: '-0.4rem',
+  large: '-0.5rem',
+  xLarge: '-0.5rem',
+  xxLarge: '-0.4rem',
+};
+
+const getVerticalAlign = (size?: TextSize) => {
+  return size ? verticalAligns[size] : undefined;
 };
 
 /**
