+++ src/components/Tooltip/Tooltip.styles.ts
@@ -1,10 +1,10 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 import { palette } from '../../util/colors';
 import { TooltipProps } from './Tooltip';
-import { borderRadiusSoft } from '../../util/styles';
+import { borderRadiusSoft } from '../../util/styles/borders';
 import { ITooltipStyles } from 'office-ui-fabric-react/lib/components/Tooltip/Tooltip.types';
 import { ICalloutContentStyles } from 'office-ui-fabric-react/lib/components/Callout/Callout.types';
-import { lineHeights, fontSizes } from '../Text/Text.styles';
+import { lineHeights, fontSizes } from '../../util/styles/fonts';
 
 export const getTooltipStyles = (_: TooltipProps): ITooltipStyles => {
   return {
