--- src/util/styles.ts
@@ -1,13 +1,0 @@
-/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-import { IRawStyle } from '@uifabric/styling';
-
-export const borderRadiusSoft = '2px';
-export const borderRadiusRound = '50%';
-export const fontWeightBold = '600';
-export const fontWeightNormal = '400';
-
-export const ellipsisStyle: IRawStyle = {
-  overflow: 'hidden',
-  textOverflow: 'ellipsis',
-  whiteSpace: 'nowrap',
-};
