+++ src/util/styles/borders.ts
@@ -1,0 +1,3 @@
+/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
+export const borderRadiusSoft = '2px';
+export const borderRadiusRound = '50%';
