+++ src/util/styles/gutters.ts
@@ -1,0 +1,12 @@
+/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
+import { GutterSize } from '../../components/FixedGrid/enums';
+
+export const gutterSize: { [size in GutterSize]: string } = {
+  none: '0',
+  xSmall: '0.4rem',
+  small: '0.8rem',
+  medium: '1.2rem',
+  large: '1.6rem',
+  xLarge: '2rem',
+  xxLarge: '2.4rem',
+};
